+++ src/main/java/net/neganote/gtutilities/common/item/UtilItems.java
@@ -4,10 +4,11 @@ import com.gregtechceu.gtceu.api.item.IComponentItem;
 import com.gregtechceu.gtceu.api.item.component.ElectricStats;
 import com.gregtechceu.gtceu.api.item.component.IItemComponent;
 
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.ItemTags;
+import net.minecraft.world.item.Item;
 import net.neganote.gtutilities.GregTechModernUtilities;
 import net.neganote.gtutilities.config.UtilConfig;
+import net.neganote.gtutilities.datagen.UtilModels;
 
 import com.tterrag.registrate.util.entry.ItemEntry;
 import com.tterrag.registrate.util.nullness.NonNullConsumer;
@@ -24,15 +25,25 @@ public class UtilItems {
     public static ItemEntry<OmniBreakerItem> OMNIBREAKER = null;
     public static int OMNIBREAKER_TIER = UtilConfig.INSTANCE.features.omnibreakerTier;
 
+    public static ItemEntry<Item> ANCIENT_GOLD_COIN = null;
+    public static ItemEntry<Item> CHOCOLATE_COIN = null;
+    public static ItemEntry<Item> COPPER_CREDIT = null;
+    public static ItemEntry<Item> CUPRONICKEL_CREDIT = null;
+    public static ItemEntry<Item> DOGE_COIN = null;
+    public static ItemEntry<Item> GOLD_CREDIT = null;
+    public static ItemEntry<Item> NAQUADAH_CREDIT = null;
+    public static ItemEntry<Item> NEUTRONIUM_CREDIT = null;
+    public static ItemEntry<Item> OSMIUM_CREDIT = null;
+    public static ItemEntry<Item> PLATINUM_CREDIT = null;
+    public static ItemEntry<Item> SILVER_CREDIT = null;
+
     static {
         if (UtilConfig.INSTANCE.features.omnibreakerEnabled) {
             OMNIBREAKER = REGISTRATE
                     .item("omnibreaker", (p) -> OmniBreakerItem.create(p, OMNIBREAKER_TIER))
                     .lang("Omni-breaker")
                     .model((ctx, prov) -> {
-                        ResourceLocation handheldItem = new ResourceLocation("item/handheld");
-                        prov
-                                .withExistingParent(ctx.getName(), handheldItem)
+                        prov.handheld(ctx)
                                 .texture("layer0", GregTechModernUtilities.id("item/omnibreaker"));
                         // .override()
                         // .predicate(GregTechModernUtilities.id("omnibreaker_name"), 1)
@@ -53,6 +64,32 @@ public class UtilItems {
                             new PrecisionBreakBehavior(OMNIBREAKER_TIER)))
                     .register();
         }
+
+        if (UtilConfig.INSTANCE.features.coinsEnabled) {
+            ANCIENT_GOLD_COIN = registerBasicItem("ancient_gold_coin");
+            CHOCOLATE_COIN = registerBasicItem("chocolate_coin");
+            COPPER_CREDIT = registerBasicItem("copper_credit");
+            CUPRONICKEL_CREDIT = registerBasicItem("cupronickel_credit");
+            DOGE_COIN = registerBasicItem("doge_coin");
+            GOLD_CREDIT = registerBasicItem("gold_credit");
+            NAQUADAH_CREDIT = registerBasicItem("naquadah_credit");
+            NEUTRONIUM_CREDIT = registerBasicItem("neutronium_credit");
+            OSMIUM_CREDIT = registerBasicItem("osmium_credit");
+            PLATINUM_CREDIT = registerBasicItem("platinum_credit");
+            SILVER_CREDIT = registerBasicItem("silver_credit");
+        }
+    }
+
+    public static ItemEntry<Item> registerBasicItem(String id, String texturePath) {
+        return REGISTRATE
+                .item(id, Item::new)
+                .initialProperties(Item.Properties::new)
+                .model(UtilModels.basicItemModel(texturePath))
+                .register();
+    }
+
+    public static ItemEntry<Item> registerBasicItem(String id) {
+        return registerBasicItem(id, id);
     }
 
     public static void init() {}